knowledge_graph/requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
colorama==0.4.4
filelock==3.6.0
grpcio==1.43.0
huggingface-hub==0.4.0
idna==3.3
joblib==1.1.0
langcodes==3.2.1
nltk==3.6.5
numpy==1.22.3
packaging==21.3
pandas==1.4.2
pathy==0.6.1
Pillow==9.1.0
protobuf==3.15.5
pydantic==1.8.2
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
regex==2022.3.15
requests==2.27.1
sacremoses==0.0.49
scikit-learn==1.0.2
scipy==1.8.0
sentence-transformers==2.2.0
sentencepiece==0.1.96
six==1.16.0
smart-open==5.2.1
spacy==3.2.0
spacy-legacy==3.0.8
spacy-loggers==1.0.1
thinc==8.0.13
threadpoolctl==3.1.0
tokenizers==0.11.6
torch==1.11.0
torchvision==0.12.0
tqdm==4.64.0
transformers==4.17.0
typedb-client==2.8.0
typedb-protocol==2.6.0
typing-extensions==3.10.0.2
ujson==5.1.0
urllib3==1.26.9
